+++ lib/ronin/support/network/tcp.rb
@@ -131,8 +131,11 @@ module Ronin
                    end
 
           if block_given?
-            yield socket
-            socket.close
+            begin
+              yield socket
+            ensure
+              socket.close
+            end
           else
             return socket
           end